+++ src/WFU.Bridge/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace WFU.Bridge;
-
-public class Class1
-{
-
-}