+++ src/WFU.Core/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace WFU.Core;
-
-public class Class1
-{
-
-}