+++ src/WFU.PluginSDK/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace WFU.PluginSDK;
-
-public class Class1
-{
-
-}